requirements.txt
amqp==2.4.2
anyjson==0.3.3
appdirs==1.4.3
Babel==2.3.4
billiard==3.5.0.2
boto==2.46.1
boto3==1.9.15
botocore==1.12.16
celery==4.2.0
cffi==1.9.1
coreapi==2.3.0
coreschema==0.0.4
croniter==0.3.15
cryptography==1.7.2
Cython==0.25.2
Django==2.2.5
django-celery-beat==1.5.0
django-celery-results==1.0.1
django-cors-headers==2.0.2
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-filter==2.2.0
django-formtools==2.0
django-modeltranslation==0.13.3
django-mptt==0.10.0
django-redis==4.8.0
django-redisboard==3.0.2
django-rest-swagger==2.1.2
django-storages==1.5.2
djangorestframework==3.10.3
docutils==0.13.1
funcy==1.7.3
geoip2==2.4.2
idna==2.2
itypes==1.1.0
Jinja2==2.9.6
jmespath==0.9.1
kombu==4.4.0
MarkupSafe==1.0
maxminddb==1.2.3
mpmath==0.19
numpy==1.12.0
openapi-codec==1.3.1
packaging==16.8
psycopg2==2.7.1
pyasn1==0.2.3
pycountry==17.1.8
pycparser==2.17
pyOpenSSL==16.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
qrcode==5.3
redis==3.2.0
requests==2.13.0
s3transfer==0.1.10
scipy==0.18.1
simpleeval==0.9.5
simplejson==3.10.0
six==1.10.0
sqlparse==0.2.3
suds-jurko==0.6
tabulate==0.7.7
ua-parser==0.7.3
ujson==1.35
uritemplate==3.0.0
user-agents==1.1.0
vine==1.1.3
XlsxWriter==0.9.6
python-logstash==0.4.6
flower==0.9.3